% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(connectionDetails, cdmDatabaseSchema, oracleTempSchema = NULL,
  outcomeDatabaseSchema, outcomeTable,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era", nestingCohortDatabaseSchema,
  nestingCohortTable, outputFolder, databaseName, maxCores = 1,
  cdmVersion = "5", createNegativeControlCohorts = TRUE,
  imputeExposureLengthForPanther = TRUE,
  synthesizePositiveControls = TRUE, runCohortMethod = TRUE,
  runSelfControlledCaseSeries = TRUE, runSelfControlledCohort = TRUE,
  runCaseControl = TRUE, runCaseCrossover = TRUE,
  packageResults = TRUE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outcomeDatabaseSchema}{Schema name where outcome data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{outcomeTable}{The name of the table that will be created in the outcomeDatabaseSchema.
This table will hold the outcome cohorts used in this
study.}

\item{exposureDatabaseSchema}{For PanTher only: Schema name where exposure data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{exposureTable}{For PanTher only: The name of the table that will be created in the exposureDatabaseSchema
This table will hold the exposure cohorts used in this
study.}

\item{nestingCohortDatabaseSchema}{Schema name where nesting cohort data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{nestingCohortTable}{The name of the table that will be created in the nestingCohortDatabaseSchema
This table will hold the nesting cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseName}{A short string for identifying the database (e.g.
'Synpuf').}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{cdmVersion}{Version of the Common Data Model used. Currently only version 5 is supported.}

\item{createNegativeControlCohorts}{Create the negative control outcome and nesting cohorts?}

\item{imputeExposureLengthForPanther}{For PanTher only: impute exposure length?}

\item{synthesizePositiveControls}{Should positive controls be synthesized?}

\item{runCohortMethod}{Perform the cohort method analyses?}

\item{runSelfControlledCaseSeries}{Perform the SCCS analyses?}

\item{runSelfControlledCohort}{Perform the SCC analyses?}

\item{runCaseControl}{Perform the case-control analyses?}

\item{runCaseCrossover}{Perform the case-crossover analyses?}

\item{packageResults}{Should results be packaged for later sharing and viewing?}
}
\description{
Execute the Study
}
\details{
This function executes the Study.

The \code{createCohorts}, \code{synthesizePositiveControls}, \code{runAnalyses}, and \code{runDiagnostics} arguments
are intended to be used to run parts of the full study at a time, but none of the parts are considerd to be optional.
}
